use custom hook comment container

+++ utils/hooks/useContent.js
@@ -5,17 +5,21 @@ export function useContent({ mutateMemorize, actionMemorize, deleteMemorize }) {
   const dispatch = useDispatch()
 
   const onLike = useCallback((id, action) => {
-    dispatch(actionMemorize({ id, action }))
+    if (actionMemorize) {
+      dispatch(actionMemorize({ id, action }))
+    }
   }, [])
 
   const onDelete = useCallback(id => {
-    dispatch(deleteMemorize(id))
+    if (deleteMemorize) {
+      dispatch(deleteMemorize(id))
+    }
   }, [])
 
   const onMemorize = useCallback(({ id, content, articleId, setContent }) => {
-    dispatch(mutateMemorize({ id, content, articleId }))
-    if (setContent) {
-      setContent('')
+    if (mutateMemorize) {
+      dispatch(mutateMemorize({ id, content, articleId }))
+      if (setContent) setContent('')
     }
   }, [])
